Move setup.py commands out of fipy package
