+++ _setup/copy_script.py
